changed Sequence and GlyphCodec get methods to take longs instead of ints.  All implementations still use ints to internally store their data (since they are array based) but this will allow us to change that in the future.
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1326 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/symbol/Sequence.java
@@ -45,7 +45,7 @@ public interface Sequence<T extends Symbol> extends Iterable<T>{
      * @return the {@link Symbol} at the specified index;
      * will never be null.
      */
-    T get(int index);
+    T get(long index);
     /**
      * Get the number of {@link Symbol}s that are in
      * this sequence.
